Refactor the code by organizing it in separate files
Extracted code from app.py and placed it into separate files to keep it more manageable.
